refactor: move deploy script to a separate package
