+++ package.json
@@ -28,8 +28,6 @@
     "npm": ">=7.0.0",
     "node": ">=14.0.0"
   },
-  "dependencies": {
-    "fs-extra": "^10.1.0"
-  },
+  "dependencies": {},
   "packageManager": "yarn@1.22.5"
 }
